// client/src/Contexts/User/ActionKind.tsx
export enum ActionKind {
  Login = "user/login",
  LoginRejected = "user/rejected",
  GetMe = "user/getMe",
  SetUser = "user/setUser",
  Registration = "user/registration",
  ChangeInformation = "user/change",
  UploadImage = "user/upload",
  SetTokens = "user/tokens",
  Loading = "cart/loading",
}